pyproject.toml
[project]
name = "learning-rag"
version = "0.1.0"
description = "Learning Retrieval-Augmented Generation (RAG)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.1.0",
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-openai>=1.1.7",
    "langchain-qdrant>=0.2",
    "langchain-text-splitters>=0.3",
    "python-dotenv>=1.0",
    "pypdf>=5.0",
    "qdrant-client>=1.16.2",
]

[tool.black]
line-length = 88
target-version = ["py313"]

